src/requirements.txt
flask
redis
mysql-connector-python
mysql-connector
PyMySQL
cryptography
